// bake/schema/openapi-types.ts
interface Map<T> {
  [key: string]: T;
}
// ...
export interface ServerObject {
  url: string;
  description?: string;
  variables?: Map<ServerVariable>;
}

export interface ServerVariable {
  enum?: string[];
  default: string;
  description?: string;
}
// ...
export type PathsObjectWithRef = Map<PathItemObjectWithRef>;

export interface PathItemObject {
  $ref?: string;
  summary?: string;
  description?: string;
  get?: OperationObject;
  put?: OperationObject;
  post?: OperationObject;
  delete?: OperationObject;
  options?: OperationObject;
  head?: OperationObject;
  patch?: OperationObject;
  trace?: OperationObject;
  servers?: ServerObject[];
  parameters?: ParameterObject[];
}

export interface PathItemObjectWithRef {
  $ref?: string;
  allOf?: ReferenceObject[];
  summary?: string;
  description?: string;
  get?: OperationObjectWithRef;
  put?: OperationObjectWithRef;
  post?: OperationObjectWithRef;
  delete?: OperationObjectWithRef;
  options?: OperationObjectWithRef;
  head?: OperationObjectWithRef;
  patch?: OperationObjectWithRef;
  trace?: OperationObjectWithRef;
  servers?: ServerObject[];
  parameters?: (ParameterObjectWithRef | ReferenceObject)[];
}

export interface OperationObject {
  tags?: string[];
  summary?: string;
  description?: string;
  externalDocs?: ExternalDocumentationObject;
  operationId?: string;
  parameters?: ParameterObject[];
  requestBody?: RequestBodyObject;
  responses: ResponsesObject;
  callbacks?: Map<CallbackObject>;
  deprecated?: boolean;
  security?: SecurityRequirementObject[];
  servers?: ServerObject[];

  // extensions
  "x-deprecated-description"?: string;
}

export interface OperationObjectWithRef {
  tags?: string[];
  summary?: string;
  description?: string;
  externalDocs?: ExternalDocumentationObject;
  operationId?: string;
  parameters?: (ParameterObjectWithRef | ReferenceObject)[];
  requestBody?: RequestBodyObjectWithRef | ReferenceObject;
  responses: ResponsesObjectWithRef;
  callbacks?: Map<CallbackObjectWithRef | ReferenceObject>;
  deprecated?: boolean;
  security?: SecurityRequirementObject[];
  servers?: ServerObject[];

  // extensions
  "x-deprecated-description"?: string;
}

export interface ExternalDocumentationObject {
  description?: string;
  url: string;
}

export interface ParameterObject {
  name: string;
  in: "query" | "header" | "path" | "cookie";
  description?: string;
  required?: boolean;
  deprecated?: boolean;
  allowEmptyValue?: boolean;
  //
  style?: string;
  explode?: string;
  allowReserved?: boolean;
  schema?: SchemaObject;
  example?: any;
  examples?: Map<ExampleObject>;
  //
  content?: Map<MediaTypeObject>;
  param?: Object;
  // ignoring stylings: matrix, label, form, simple, spaceDelimited,
  // pipeDelimited and deepObject
}

export interface ParameterObjectWithRef {
  name: string;
  in: string;
  description?: string;
  required?: boolean;
  deprecated?: boolean;
  allowEmptyValue?: boolean;
  //
  style?: string;
  explode?: string;
  allowReserved?: boolean;
  schema?: SchemaObjectWithRef | ReferenceObject;
  example?: any;
  examples?: Map<ExampleObject | ReferenceObject>;
  //
  content?: Map<MediaTypeObjectWithRef>;
  // ignoring stylings: matrix, label, form, simple, spaceDelimited,
  // pipeDelimited and deepObject
}

export interface RequestBodyObject {
  description?: string;
  content: Map<MediaTypeObject>;
  required?: boolean;
}

export interface RequestBodyObjectWithRef {
  description?: string;
  content: Map<MediaTypeObjectWithRef>;
  required?: boolean;
}

export interface MediaTypeObject {
  schema?: SchemaObject;
  example?: any;
  examples?: Map<ExampleObject>;
  encoding?: Map<EncodingObject>;
}

export interface MediaTypeObjectWithRef {
  schema?: SchemaObjectWithRef | ReferenceObject;
  example?: any;
  examples?: Map<ExampleObject | ReferenceObject>;
  encoding?: Map<EncodingObjectWithRef>;
}

export interface EncodingObject {
  contentType?: string;
  headers?: Map<HeaderObject>;
  style?: string;
  explode?: boolean;
  allowReserved?: boolean;
}

export interface EncodingObjectWithRef {
  contentType?: string;
  headers?: Map<HeaderObjectWithRef | ReferenceObject>;
  style?: string;
  explode?: boolean;
  allowReserved?: boolean;
}

export type ResponsesObject = Map<ResponseObject>;

export type ResponsesObjectWithRef = Map<
  ResponseObjectWithRef | ReferenceObject
>;

export interface ResponseObject {
  description: string;
  headers?: Map<HeaderObject>;
  content?: Map<MediaTypeObject>;
  links?: Map<LinkObject>;
}

export interface ResponseObjectWithRef {
  description: string;
  headers?: Map<HeaderObjectWithRef | ReferenceObject>;
  content?: Map<MediaTypeObjectWithRef>;
  links?: Map<LinkObject | ReferenceObject>;
}

export type CallbackObject = Map<PathItemObject>;

export type CallbackObjectWithRef = Map<PathItemObjectWithRef>;

export interface ExampleObject {
  summary?: string;
  description?: string;
  value?: any;
  externalValue?: string;
}

export interface LinkObject {
  operationRef?: string;
  operationId?: string;
  parameters?: Map<any>;
  requestBody?: any;
  description?: string;
  server?: ServerObject;
}

export type HeaderObject = Omit<ParameterObject, "name" | "in">;

export type HeaderObjectWithRef = Omit<ParameterObjectWithRef, "name" | "in">;

export interface TagObject {
  name?: string;
  description?: string;
  externalDocs?: ExternalDocumentationObject;
  "x-displayName"?: string;
}

export interface ReferenceObject {
  $ref: string;
}

export type JSONSchema = JSONSchema4 | JSONSchema6 | JSONSchema7;
export type SchemaObject =
  & Omit<
    JSONSchema,
    | "type"
    | "allOf"
    | "oneOf"
    | "anyOf"
    | "not"
    | "items"
    | "properties"
    | "additionalProperties"
  >
  & {
    // OpenAPI specific overrides
    type?: "string" | "number" | "integer" | "boolean" | "object" | "array";
    allOf?: SchemaObject[];
    oneOf?: SchemaObject[];
    anyOf?: SchemaObject[];
    not?: SchemaObject;
    items?: SchemaObject;
    properties?: Map<SchemaObject>;
    additionalProperties?: Map<SchemaObject>;

    // OpenAPI additions
    nullable?: boolean;
    discriminator?: DiscriminatorObject;
    readOnly?: boolean;
    writeOnly?: boolean;
    xml?: XMLObject;
    externalDocs?: ExternalDocumentationObject;
    example?: any;
    deprecated?: boolean;

    // For ref tracking
    "#ref"?: string;
  };

export type SchemaObjectWithRef =
  & Omit<
    JSONSchema,
    | "type"
    | "allOf"
    | "oneOf"
    | "anyOf"
    | "not"
    | "items"
    | "properties"
    | "additionalProperties"
  >
  & {
    // OpenAPI specific overrides
    type?: "string" | "number" | "integer" | "boolean" | "object" | "array";
    allOf?: (SchemaObjectWithRef | ReferenceObject)[];
    oneOf?: (SchemaObjectWithRef | ReferenceObject)[];
    anyOf?: (SchemaObjectWithRef | ReferenceObject)[];
    not?: SchemaObjectWithRef | ReferenceObject;
    items?: SchemaObjectWithRef | ReferenceObject;
    properties?: Map<SchemaObjectWithRef | ReferenceObject>;
    additionalProperties?: boolean | SchemaObjectWithRef | ReferenceObject;

    // OpenAPI additions
    nullable?: boolean;
    discriminator?: DiscriminatorObject;
    readOnly?: boolean;
    writeOnly?: boolean;
    xml?: XMLObject;
    externalDocs?: ExternalDocumentationObject;
    example?: any;
    deprecated?: boolean;

    // For ref tracking
    "#ref"?: string;
  };

export interface DiscriminatorObject {
  propertyName: string;
  mapping?: Map<string>;
}

export interface XMLObject {
  name?: string;
  namespace?: string;
  prefix?: string;
  attribute?: boolean;
  wrapped?: boolean;
}
// ...
export type SecurityRequirementObject = Map<string[]>;